Add many patches, and modify build.sh
build.sh now applies patches using 'git am' instead of patch.  A bit
slower, but it preserves info about the developers who created the
patches.

It also now applies the patches in the order specified in the file
patch-order.txt.  There are some pairs of patches that apply cleanly
in one order but not the opposite order.